+++ src/errors/fullscreen-error/slots/index.ts
@@ -1,0 +1,1 @@
+export * from './error-header-slot';