+++ src/errors/fullscreen-error/styles.ts
@@ -1,0 +1,14 @@
+import styled from '@emotion/styled';
+
+export const StyledScreenError = styled('div')`
+	width: 100%;
+	height: 100%;
+
+	display: flex;
+	flex-direction: column;
+
+	justify-content: center;
+	align-items: center;
+
+	gap: 12px;
+`;