+++ src/index.ts
@@ -13,3 +13,4 @@ export * from './card';
 export * from './hoverable';
 export * from './menu';
 export * from './avatar';
+export * from './errors';