+++ src/layout/index.ts
@@ -1,2 +1,3 @@
 export * from './layout-full-window';
 export * from './page-layout';
+export * from './default-layout';