+++ src/layout/page-layout/components/header/styles.ts
@@ -4,4 +4,19 @@ export const PageHeaderWrapper = styled('div')`
 	width: 100%;
 
 	padding: 12px 12px 12px 0px;
+
+	display: flex;
+
+	justify-content: space-between;
+
+	align-items: center;
+`;
+
+export const PageHeaderActions = styled('div')`
+	width: fit-content;
+
+	gap: 12px;
+
+	display: flex;
+	flex-direction: row;
 `;