+++ src/stories/Error.stories.tsx
@@ -1,0 +1,36 @@
+import React from 'react';
+import type { Meta, StoryObj } from '@storybook/react';
+import { FullscreenError } from '../errors';
+import { ErrorIllustration } from '../assets/error-illustration';
+
+// More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
+const meta = {
+	title: 'Components/Errors',
+	component: FullscreenError,
+	parameters: {
+		// Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
+		layout: 'centered',
+	},
+	// This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/writing-docs/autodocs
+	tags: ['autodocs'],
+	argTypes: {},
+	// More on argTypes: https://storybook.js.org/docs/api/argtypes
+	// Use `fn` to spy on the onClick arg, which will appear in the actions panel once invoked: https://storybook.js.org/docs/essentials/actions#action-args
+	args: {
+		image: 'https://acropora.ru/images/yootheme/pages/features/panel03.jpg',
+		title: 'Ошибка',
+		description: 'Фатальная ошибка',
+	},
+} satisfies Meta<typeof FullscreenError>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+// More on writing stories with args: https://storybook.js.org/docs/writing-stories/args
+export const Primary: Story = {
+	args: {
+		image: <ErrorIllustration />,
+		title: 'Ошибка',
+		description: 'Фатальная ошибка',
+	},
+};